Refactored some work to clean things up
